+++ src/main/kotlin/me/nobaboy/nobaaddons/features/events/mythological/GriffinBurrowGuess.kt
@@ -2,6 +2,9 @@ package me.nobaboy.nobaaddons.features.events.mythological
 
 import me.nobaboy.nobaaddons.api.skyblock.events.mythological.DianaAPI
 import me.nobaboy.nobaaddons.config.NobaConfig
+import me.nobaboy.nobaaddons.events.EventDispatcher.Companion.registerIf
+import me.nobaboy.nobaaddons.events.impl.interact.BlockInteractionEvent
+import me.nobaboy.nobaaddons.events.impl.interact.GenericInteractEvent
 import me.nobaboy.nobaaddons.events.impl.interact.ItemUseEvent
 import me.nobaboy.nobaaddons.events.impl.render.ParticleEvents
 import me.nobaboy.nobaaddons.events.impl.skyblock.MythologicalEvents
@@ -23,6 +26,7 @@ object GriffinBurrowGuess {
 		SkyBlockEvents.ISLAND_CHANGE.register { reset() }
 		ParticleEvents.PARTICLE.register(this::onParticle)
 		ItemUseEvent.EVENT.register(this::onItemUse)
+		BlockInteractionEvent.EVENT.registerIf<BlockInteractionEvent.Interact>(this::onItemUse)
 	}
 
 	private fun onParticle(event: ParticleEvents.Particle) {
@@ -41,12 +45,10 @@ object GriffinBurrowGuess {
 		MythologicalEvents.BURROW_GUESS.dispatch(MythologicalEvents.BurrowGuess(guessLocation.lower(0.5).roundToBlock()))
 	}
 
-	private fun onItemUse(event: ItemUseEvent) {
+	private fun onItemUse(event: GenericInteractEvent) {
 		if(!enabled) return
 		if(lastAbilityUse.elapsedSince() <= 3.seconds) return
-
-		val itemId = event.itemInHand.skyBlockId ?: return
-		if(itemId != DianaAPI.SPADE) return
+		if(event.itemInHand.skyBlockId != DianaAPI.SPADE) return
 
 		particlePath.reset()
 		lastAbilityUse = Timestamp.now()